Add Temporal MV scan support and fixes in decoder related to parsing, ALTREF, monochrome trans_info update

+++ Source/Lib/Common/Codec/EbAv1Structs.h
@@ -373,7 +373,7 @@ typedef struct LRParams {
 
     /*!< Specifies the size of loop restoration units in units of samples in
      * the current plane */
-    uint8_t             loop_restoration_size;
+    uint16_t            loop_restoration_size;
 
 } LRParams;
